+++ django_dramatiq/middleware.py
@@ -3,6 +3,8 @@ import logging
 from django import db
 from dramatiq.middleware import Middleware
 
+from .models import Task
+
 LOGGER = logging.getLogger("django_dramatiq.AdminMiddleware")
 
 
@@ -11,7 +13,6 @@ class AdminMiddleware(Middleware):
     """
 
     def after_enqueue(self, broker, message, delay):
-        from .models import Task
         LOGGER.debug("Creating Task from message %r.", message.message_id)
         status = Task.STATUS_ENQUEUED
         if delay:
@@ -20,12 +21,10 @@ class AdminMiddleware(Middleware):
         Task.tasks.create_or_update_from_message(message, status=status)
 
     def before_process_message(self, broker, message):
-        from .models import Task
         LOGGER.debug("Updating Task from message %r.", message.message_id)
         Task.tasks.create_or_update_from_message(message, status=Task.STATUS_RUNNING)
 
     def after_process_message(self, broker, message, *, result=None, exception=None):
-        from .models import Task
         status = Task.STATUS_DONE
         if exception is not None:
             status = Task.STATUS_FAILED